Move createInternalColumn() as a follow up to #4994
